refactor(1013): extract CLIShellManager facade (Cat B, position 30) (#836)
Extract CLIShellManager from MistHelper.py to src/ssh/cli_shell_manager.py
using the SC-001 fresh-extraction facade pattern.

- New module: src/ssh/cli_shell_manager.py with lazy mh.PromptClientUtils,
  mh.apisession, and mh.KeyboardListener accessors inside helpers that need
  live MistHelper globals; stdlib and installed packages (mistapi, websocket,
  optional pyte) are direct-imported.
- MistHelper.py: alphabetized re-export alias inserted, class body replaced
  with a 2-line breadcrumb, stale shutil import removed, KeyboardListener
  import kept with noqa: F401 since the extracted module resolves it via
  mh.KeyboardListener.
- Cleans up the pre-existing duplicate @staticmethod decorator bug on
  _shell_start_receiver during the extraction.
- pyproject.toml: coverage-omit entry added.

Local gate: black + ruff clean; py_compile OK; --test 66/0/131 baseline
holds.
