+++ pyproject.toml
@@ -248,6 +248,7 @@ omit = [
     "src/reports/sfp_transceiver_data_processor.py",
     "src/reports/wired_client_manufacturer_report_generator.py",
     "src/site/bulk_radius_wlan_config_manager.py",
+    "src/ssh/cli_shell_manager.py",
     "src/ui/tui.py",
     "src/websocket/*"
 ]
